apps/api-core-v2/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
alembic==1.14.1
SQLAlchemy==2.0.38
asyncpg==0.30.0
psycopg[binary]==3.2.5

# Auth dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.20
email-validator==2.2.0

# Celery and Redis for background jobs
celery==5.4.0
redis==5.2.1

# Redis async for caching
redis[asyncio]==5.2.1

# Stripe for escrow payments
stripe==11.1.1

# Rate limiting
slowapi==0.1.9

# Structured logging
structlog==24.1.0

# Sentry for Error Tracking
sentry-sdk[fastapi]==2.8.0

